+++ README.md
@@ -1,6 +1,10 @@
 # PDO-Library
-Simple PHP library with PDO and PHP 5.6.0. To load files automatically PSR 4 autoloading is used.
+Simple PHP Mini Framework with PDO and PHP 5.6.0. 
+-- To load files automatically PSR 4 autoloading is used.
+-- For Templating Twig template engine.
+-- For Debugging Kint debugger.
+-- For SMTP mails phpMailer.
 
-To use this library, you need to only include config.php file from 'app' directory.
+To use this library, you need to only include Bootstrap.php file from 'app' directory.
 
-example : require_nce "app/config.php";
+example : require_once "app/Bootstrap.php";
