+++ composer.json
@@ -6,7 +6,10 @@
     "require": {
         "php": ">=5.4"
     },
-    "autoload": {        
+    "autoload": { 
+        "files": [
+            "app/System/Functions.php"
+        ],       
         "psr-4": {
             "App\\": "app/"
         }
